requirements_2.txt
acme==1.28.0
appdirs==1.4.4
asgiref==3.5.2
attrs==21.4.0
btrfsutil==5.18.1
CacheControl==0.12.6
certbot==1.28.0
certbot-nginx==1.28.0
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
ConfigArgParse==1.5
configobj==5.1.0.dev0
contextlib2==21.6.0
cryptography==37.0.2
distlib==0.3.4
distro==1.7.0
Django==4.0.6
django-cors-headers==3.13.0
django-haystack==3.2.1
django-mysql==4.7.0
django-taggit==3.0.0
django-tastypie==0.14.4
docutils==0.18.1
elastic-transport==8.1.2
elasticsearch==8.3.1
filelock==3.5.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
Jinja2==3.1.2
josepy==1.13.0
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
lit==13.0.1.dev0
MarkupSafe==2.1.1
mock==3.0.5
more-itertools==8.13.0
msgpack==1.0.3
mysqlclient==2.1.1
netifaces==0.11.0
nspektr==0.4.0
oauthlib==3.2.0
ordered-set==4.0.2
packaging==21.3
parsedatetime==2.6
pbr==5.9.0
pep517==0.12.0
platformdirs==2.5.1
ply==3.11
progress==1.6
pycparser==2.21
Pygments==2.12.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.1
PyYAML==6.0
Reflector==2021.11.20.2.41.3
requests==2.27.1
requests-toolbelt==0.9.1
resolvelib==0.8.1
retrying==1.3.3
rich @ file:///build/python-rich/src/rich-12.4.4/dist/rich-12.4.4-py3-none-any.whl
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
trove-classifiers==2022.6.26
typing_extensions==4.2.0
ufw==0.36.1
urllib3==1.26.9
validate==5.1.0.dev0
validate-pyproject==0.9
virtualenv==20.11.0
webencodings==0.5.1
zipp==3.8.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.proxy==4.5.0
